env: Add support for loading and importing plain text env files
Using boot scripts have the advantage that they use U-Boot
binary format which means that they have a header with checksum,
timestamp, etc but unfortunate this means that whatever updates
a Linux distribution has to call mkimage to build the bootscript.

Another option is to use U-Boot's env import command to import a
set of variables from a plain text file. U-Boot expect a set of
standard variables to be defined in this file like ${kernel_image},
${ramdisk_image} and ${bootargs}.

Add some standard commands that handle this to be used by boards.

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>